Blip_Buffer: eliminate float from volume, rate and eq path
Convert the Blip_Buffer volume chain to Q30 fixed-point and the resampling
factor to exact 64-bit integer, and make the baked kernels the only code path
in the default build (the floating-point sinc is compiled only when
regenerating blip_kernels.h). No transcendental or float math is emitted for
the audio path, so every coefficient is deterministic across compilers, libm
versions and optimization levels.

- volume_unit() -> volume_unit_fixed(): delta_factor and the attenuation
  shift are computed in 64-bit integer from a Q30 unit. Blip_Synth::volume()
  takes a Q30 scalar; every chip passes an exact round(volume*2^30) constant.
  All 13 synth delta_factors are bit-identical to the former float path.
- clock_rate_factor(): round(sample_rate/clock_rate * 2^ACCURACY) as an exact
  integer division (verified identical to the former double result).
- blip_eq_t::treble and Nes_Emu::equalizer_t::treble are now int (all presets
  are whole-dB); the kernel lookup keys on int. set_frame_rate() is integer.
- gen_sinc()/blip_eq_t::generate() guarded under BLIP_REGEN_KERNELS;
  treble_eq() loads a baked kernel with a canonical fallback, so it can never
  reach float in the default build.

Validated: float-ref vs baked kernels bit-identical across all 48 combos
(including the volume kernel-shift path); regen is byte-reproducible.

+++ nes_emu/Nes_Emu.h
@@ -115,14 +115,14 @@ public:
 
 	// Adjust effective frame rate by changing how many samples are generated each frame.
 	// Allows fine tuning of frame rate to improve synchronization.
-	void set_frame_rate( double rate );
+	void set_frame_rate( int rate );
 
 	// Number of sound channels for current cartridge
 	int channel_count() const { return channel_count_; }
 
 	// Frequency equalizer parameters
 	struct equalizer_t {
-		double treble; // 5.0 = extra-crisp, -200.0 = muffled
+		int treble; // 5 = extra-crisp, -200 = muffled (integer dB)
 		long bass;     // 0 = deep, 20000 = tinny
 	};
